ecommerce_backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.8
django-crontab==0.7.1
djangorestframework==3.16.1
docker==7.1.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
idna==3.11
promise==2.3
python-dateutil==2.9.0.post0
pywin32==311
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
